+++ demo/site/src/app/[lang]/news/[slug]/page.tsx
@@ -5,6 +5,7 @@ import { GQLNewsContentScopeInput } from "@src/graphql.generated";
 import { NewsContentBlock } from "@src/news/blocks/NewsContentBlock";
 import { createGraphQLFetch } from "@src/util/graphQLClient";
 import { notFound } from "next/navigation";
+import { FormattedDate } from "react-intl";
 
 import { GQLNewsDetailPageQuery, GQLNewsDetailPageQueryVariables } from "./page.generated";
 
@@ -37,7 +38,9 @@ export default async function NewsDetailPage({ params }: { params: { slug: strin
         <div>
             <DamImageBlock data={news.image} layout="responsive" sizes="100vw" aspectRatio="16x9" />
             <h1>{news.title}</h1>
-            {/*<p><FormattedDate value={news.createdAt} /></p>*/}
+            <p>
+                <FormattedDate value={news.createdAt} />
+            </p>
             <hr />
             <NewsContentBlock data={news.content} />
         </div>