+++ demo/site/src/blocks/FullWidthImageBlock.tsx
@@ -11,7 +11,7 @@ export const FullWidthImageBlock = withPreview(
     ({ data: { image, content } }: PropsWithData<FullWidthImageBlockData>) => {
         return (
             <Root>
-                <DamImageBlock data={image} layout="responsive" sizes="100vw" />
+                <DamImageBlock data={image} layout="responsive" sizes="100vw" aspectRatio="16x9" />
                 <OptionalBlock
                     block={(props) => (
                         <Content>